+++ src/Helper/URLSearchParams.php
@@ -1,0 +1,18 @@
+<?php
+
+// AUTOMATICALLY GENERATED.  DO NOT EDIT.
+// Use `composer build` to regenerate.
+
+namespace Wikimedia\IDLeDOM\Helper;
+
+trait URLSearchParams {
+	/**
+	 * @return string
+	 */
+	public function __toString() : string {
+		'@phan-var \Wikimedia\IDLeDOM\URLSearchParams $this';
+		// @var \Wikimedia\IDLeDOM\URLSearchParams $this
+		return $this->toString();
+	}
+
+}